Authentication

+++ app/Http/Controllers/TasksController.php
@@ -15,9 +15,17 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks
+            ];
+        }
         
-        return view('tasks.index', ['tasks' => $tasks]);
+        return view('tasks.index', $data);
     }
 
     /**
@@ -46,12 +54,11 @@ class TasksController extends Controller
             'status' => 'required|max:10'
         ]);
         
-        
-        $task = new Task();
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
-        
+        $request->user()->tasks()->create([
+            'content'=>$request->content,
+            'status'=>$request->status
+        ]);
+
         return redirect('/');
         
     }
@@ -100,9 +107,12 @@ class TasksController extends Controller
         ]);
 
         $task = Task::findOrFail($id);
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+
+        if (\Auth::id() == $task->user_id) {
+            $task->content = $request->content;
+            $task->status = $request->status;
+            $task->save();
+        }
         
         return redirect('/');
         
@@ -117,7 +127,10 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::findOrFail($id);
-        $task->delete();
+        
+        if (\Auth::id() == $task->user_id) {
+            $task->delete();
+        }
 
         return redirect('/');
     }
